Microservice/TARGET_SCRAPER/pyproject.toml
[project]
name = "fakesniper"
version = "0.1.0"
description = ""
authors = [
    {name = "Erik Badalyan",email = "Erik.Badalyan@Livanova.com"}
]
requires-python = ">=3.12"
dependencies = [
    "selenium (>=4.34.2,<5.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "openai==0.28",
    "boto3 (>=1.39.11,<2.0.0)",
    "mysql-connector-python",
    "schedule (>=1.2.2,<2.0.0)"
]

packages = [{include = "service"}]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
